Add files via upload
Why should we first document unclean issues we observe, rather than just write what we need to do to fix the issues? When your data's issues get complicated, writing how to fix each can get confusing, lengthy, and time-consuming. It can get overwhelming trying to think of how to clean something complicated immediately after documenting it.

If you are separating the assessing and cleaning steps of data wrangling, as we are in this lesson, writing only observations as a first step is good practice.

If you choose to assess an issue then immediately clean that issue (which is very much allowed), you can skip the observation and go straight to defining how to clean it (which is part of the Define-Code-Test cleaning framework you’ll see in Lesson 4).

Note: Visualizing your data (i.e., creating plots) is part of Programmatic rather than Visual Assessment. Tricky! This is because plotting data requires coding, or programming.

In this Lesson
You're going to start with visual assessment in the first part of this lesson to identify data quality issues. You'll then use programmatic assessment to identify some more data quality issues.

Toward the end of the lesson, you'll use visual (first) and programmatic assessment (second) to identify tidiness issues.